refactor(nlp): port entity extraction, clustering, and spike cores to shared/ (#5697)
Move the pure NLP logic (entity registry + matcher, tokenize/jaccard,
Jaccard news clustering, CVE/APT/FIN/leader extraction, spike decision
math) into dependency-free shared/ ESM modules so server-side MCP tools
share the exact client implementations. src modules re-export unchanged;
entity extraction characterized byte-identical before/after.
